完成文章相關API

+++ app/Http/Controllers/SerieController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Serie;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Models\SerieHasArticle;
 
@@ -13,13 +14,38 @@ class SerieController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        return response()->json([
-            'result' => Serie::with(['article' => function ($query) {
-                return  $query->with('articleDetil');
-            }, 'Categorie'])->get()
-        ]);
+
+        $pagination = json_decode($request['pagination'], true);
+        $where      = json_decode($request['where'], true);
+        $order      = json_decode($request['order'], true);
+        if ($request->has(['pagination'])) {
+            $Serie = Serie::with(['categorie','article'=>function($query){
+                return $query->with('articleDetil');
+            }]);
+            if ($where) {
+                foreach ($where as $key => $value) {
+                    $Serie->where(Str::snake($key), $value);
+                }
+            }
+            if ($order) {
+                foreach ($order as $key => $value) {
+                    $Serie->orderBy(Str::snake($key), $value);
+                }
+            }
+            $Serie = $Serie->paginate($pagination['size'], ['*'], 'page', $pagination['page']);
+
+            return response()->json([
+                'list'       => $Serie->items(),
+                'pagination' => [
+                    'page'  => $Serie->currentPage(),
+                    'size'  => $pagination['size'],
+                    'total' => $Serie->total(),
+                ]
+            ]);
+        }
+        return ['result'=> Serie::get()];
     }
 
     /**
@@ -44,10 +70,43 @@ class SerieController extends Controller
      * @param  \App\Models\Serie  $serie
      * @return \Illuminate\Http\Response
      */
-    public function show(Serie $serie)
+    public function show(Request $request, Serie $serie)
     {
+        $pagination = json_decode($request['pagination'], true);
+        $where      = json_decode($request['where'], true);
+        $order      = json_decode($request['order'], true);
+
+        $article = $serie->article()->with(['articleDetil']);
+        if ($where) {
+            foreach ($where as $key => $value) {
+                $article->where(Str::snake($key), $value);
+            }
+        }
+        if ($order) {
+            foreach ($order as $key => $value) {
+                $article->orderBy(Str::snake($key), $value);
+            }
+        }
+        $article = $article->paginate($pagination['size'], ['*'], 'page', $pagination['page']);
+
         return response()->json([
-            'result' => $serie
+            'serieData'      => $serie->load(['serieHasArticle']),
+            'article' => [
+                'list'       => $article->items(),
+                'pagination' => [
+                    'page'  => $article->currentPage(),
+                    'size'  => $pagination['size'],
+                    'total' => $article->total(),
+                ]
+            ],
+            'article' => [
+                'list'       => $article->items(),
+                'pagination' => [
+                    'page'  => $article->currentPage(),
+                    'size'  => $pagination['size'],
+                    'total' => $article->total(),
+                ]
+            ]
         ]);
     }
     /**
@@ -59,12 +118,13 @@ class SerieController extends Controller
      */
     public function update(Request $request, Serie $serie)
     {
-        SerieHasArticle::where('serie_id', $serie->id)->delete();
-        foreach ($request['article'] as $key => $value) {
-            $data = ['article_id' => $value, 'page' => $key];
-            $serie->serieHasArticle()->create($data);
+        if ($request->has('article')) {
+            SerieHasArticle::where('serie_id', $serie->id)->delete();
+            foreach ($request['article'] as $key => $value) {
+                $data = ['article_id' => $value, 'page' => $key];
+                $serie->serieHasArticle()->create($data);
+            }
         }
-
         if ($serie->update($request->only(['categorie_id', 'name', 'description']))) {
             return response()->json([
                 'message' => '已成功更新系列區塊',
